fixing blog
