+++ blog/blog/public/index.html
@@ -64,17 +64,8 @@
             </div>
           </div>
       </div> -->
-      <div class="row">
-        <div class="col-lg-2" style="overflow-y: auto; max-height: 500px;" id="div_List">
-          <!-- Blog list will automatically inserted here  -->
-          <ol class="breadcrumb blogItems" onclick="blog_click(this)" id="blog_no_0">
-            <li class="breadcrumb-item active">God Mode</li>
-          </ol>
-        </div>
-        <div class="col-lg-10" id="div_content">
-
-        </div>
-        <div id="root"></div>
+      <div id="div_content">
+        
         <!--
           This HTML file is a template.
           If you open it directly in the browser, you will see an empty page.
